CP-13974: Fix Ledger auto-import toast timing and stale walletId
Move onWalletImported dispatch out of AppConnectionOnboardingScreen and
into the route-level callbacks. For onboarding, defer it to the
confirmation screen (right before login) so the toast doesn't appear
mid-onboarding. For signed-in import, pass walletId directly from
createLedgerWallet's return value via the callback to avoid a stale
closure capturing the previously active (mnemonic) wallet ID.
